fix(analytics): address review — eight findings
The three real bugs (all in the compute):
- First-touch attachment scanned EVERY accumulated episode per alert, making
  buildEpisodes quadratic. Episodes now collect per alert and a single
  forward cursor over the (sorted) events keeps attachment linear.
- An unresolved episode's span ran to infinity, so an event during the NEXT
  episode could attribute to the earlier one. Spans are now bounded by the
  next firing; regression test pins it.
- lastSeen used Math.max(...spread) — a name with tens of thousands of
  episodes would blow the argument-count limit. Now a reduce.

Scale + validation:
- The two full-table scans (history rows, event times) moved behind
  SnapshotCaches with the same TTL and invalidation as the alert snapshots —
  no more per-request scans.
- Params parse through a zod schema (AlertAnalyticsParamsSchema) that also
  rejects an inverted from/to window.

Client:
- Bounded presets re-anchor to "now" on a 5-minute bucket, so a page left
  open doesn't quietly serve yesterday's window; the bucket doubles as the
  query key so the cache doesn't churn per render.
- Loading/error/data render as a single exclusive chain.
- Charts split one component per file under Analytics/charts/ with a shared
  chartTheme, per the house convention.

Also: API-level tests for the endpoint (200 contract, 400 on inverted or
malformed windows, filter scoping) — added after this round exposed that
unit-only coverage let an unimported schema symbol slip through lint but not
tests. Full suites: 252 server, 224 client.
