+++ apps/client/src/components/Analytics/charts/index.ts
@@ -1,0 +1,6 @@
+export { HourBarChart } from './HourBarChart';
+export { MttrTrendChart } from './MttrTrendChart';
+export { SeverityDonut } from './SeverityDonut';
+export { TopList } from './TopList';
+export { VolumeAreaChart } from './VolumeAreaChart';
+export { WeekdayBarChart } from './WeekdayBarChart';
